@@ -1,7 +1,7 @@
 module.exports = {
     "env": {
         "browser": true,
-        "es2021": true
+        "es2022": true
     },
     "extends": [
         "eslint:recommended",
@@ -14,10 +14,7 @@ module.exports = {
             },
             "files": [
                 ".eslintrc.{js,cjs}"
-            ],
-            "parserOptions": {
-                "sourceType": "script"
-            }
+            ]
         }
     ],
     "parserOptions": {
