@@ -2,72 +2,7 @@
   <v-app>
     <MainNavbar />
     <v-container style="margin-top: 100px; max-width: 1000px">
-      <div style="margin-bottom: 200px;">
-        <div class="events-title">
-          Calendar
-        </div>
-        <div style="height: 600px;">
-          <!-- Calendar Navigation -->
-          <v-sheet tile height="54" class="d-flex">
-            <v-btn icon class="ma-2" @click="$refs.calendar.prev()">
-              <v-icon>mdi-chevron-left</v-icon>
-            </v-btn>
-            <v-spacer />
-            <div class="events-calendar-title" v-if="$refs.calendar">
-              {{ $refs.calendar.title }}
-            </div>
-            <v-spacer />
-            <v-select
-              v-model="calendarType"
-              :items="calendarTypes"
-              dense
-              outlined
-              hide-details
-              class="ma-2"
-              label="type"
-            ></v-select>
-            <v-btn icon class="ma-2" @click="$refs.calendar.next()">
-              <v-icon>mdi-chevron-right</v-icon>
-            </v-btn>
-          </v-sheet>
-          <!-- Calendar -->
-          <v-calendar
-            ref="calendar"
-            v-model="value"
-            :type="calendarType"
-            :events="events"
-            :event-overlap-mode="mode"
-            :event-overlap-threshold="30"
-            :event-color="getEventColor"
-            @change="getEvents"
-            @click:event="showEvent"
-          />
-          <!-- Below is code for popup on click -->
-          <v-menu
-            v-model="selectedOpen"
-            :close-on-content-click="false"
-            :activator="selectedElement"
-            offset-x
-          >
-            <v-card color="grey lighten-4" min-width="350px" flat>
-              <v-toolbar :color="selectedEvent.color" dark>
-                <!-- eslint-disable -->
-                <v-toolbar-title v-html="selectedEvent.name"></v-toolbar-title> 
-                <!-- eslint-enable -->
-                <v-spacer></v-spacer>
-                <v-btn icon @click="selectedOpen = false">
-                  <v-icon>mdi-close</v-icon>
-                </v-btn>
-              </v-toolbar>
-              <v-card-text v-if="selectedEvent.details">
-                <span v-html="selectedEvent.details"></span>
-              </v-card-text>
-            </v-card>
-          </v-menu>
-        </div>
-      </div>
-      <!-- List of events -->
-      <!-- <EventList /> -->
+      <CalendarComponent />
     </v-container>
     <MainFooter />
   </v-app>
@@ -77,6 +12,7 @@
 import "../assets/scss/events-media.scss";
 import MainNavbar from "@/layout/MainNavbar.vue";
 import MainFooter from "@/layout/MainFooter.vue";
+import CalendarComponent from "@/components/CalendarComponent.vue";
 // import EventList from "@/components/EventList";
 export default {
   name: "CalendarPage",
@@ -84,6 +20,7 @@ export default {
   components: {
     MainNavbar,
     MainFooter,
+    CalendarComponent
     // EventList,
   },
 
